+++ pages/bucket/[bucket]/collection/[collection]/record/[record].js
@@ -2,23 +2,32 @@ import React from "react";
 import Head from "next/head";
 
 import EditRecord from "../../../../../../src/EditRecord";
+import Layout from "../../../../../../src/Layout";
+
+import client from "../../../../../../src/kinto/client";
 
 const RecordPage = props => {
   return (
     <div>
       <Head>
-        <title>Dataset: {props.record}</title>
+        <title>Dataset: {props.record.id}</title>
       </Head>
-      <EditRecord {...props} />
+      <Layout>
+        <EditRecord {...props} />
+      </Layout>
     </div>
   );
 };
 
 RecordPage.getInitialProps = async ({ query }) => {
+  const record = await client
+    .bucket(query.bucket, { headers: {} })
+    .collection(query.collection, { headers: {} })
+    .getRecord(query.record, { headers: {} });
+
   return {
-    bucket: query.bucket,
-    collection: query.collection,
-    record: query.record
+    record,
+    query
   };
 };
