+++ src/forms/Glossaire.js
@@ -49,7 +49,7 @@ const StyledForm = styled(Form)`
 
 const GlossaireForm = ({ data, onSubmit, onDelete }) => (
   <React.Fragment>
-    <h1 style={{ margin: "1em 0" }}>Thèmes</h1>
+    <h1 style={{ margin: "1em 0" }}>Glossaire</h1>
     <Container>
       <Formik
         initialValues={data}
